requirements.txt
SpeechRecognition==3.8.1
pyttsx3==2.90
wikipedia==1.4.0
pywhatkit==5.4
python-dotenv==0.20.0